+++ src/lib/models/reflections/document.ts
@@ -74,6 +74,7 @@ export class DocumentReflection extends Reflection {
             content: Comment.serializeDisplayParts(serializer, this.content),
             frontmatter: this.frontmatter,
             relevanceBoost: this.relevanceBoost,
+            children: serializer.toObjectsOptional(this.children),
         };
     }
 
@@ -82,5 +83,8 @@ export class DocumentReflection extends Reflection {
         this.content = Comment.deserializeDisplayParts(de, obj.content);
         this.frontmatter = obj.frontmatter;
         this.relevanceBoost = obj.relevanceBoost;
+        this.children = de.reviveMany(obj.children, (obj) =>
+            de.reflectionBuilders.document(this, obj),
+        );
     }
 }